requirements.txt
amqp==5.1.1
argh==0.26.2
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.24.5
botocore==1.27.5
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-allauth==0.51.0
django-baton==2.3.0
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
django-filter==21.1
django-ipware==4.0.2
django-nine==0.2.5
django-redis==5.2.0
django-rest-auth-forked==0.9.5
django-storages==1.12.3
django-stubs==1.11.0
django-stubs-ext==0.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.6.0
docker-py==1.9.0
drf-yasg==1.20.0
elasticsearch==7.16.3
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
hgicommon==1.2.0
hgijson==1.3.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
kombu==5.2.4
lxml==4.9.0
MarkupSafe==2.1.1
mypy==0.950
mypy-extensions==0.4.3
mysqlclient==2.1.0
Naked==0.1.31
oauthlib==3.2.0
packaging==21.3
pathtools==0.1.2
prompt-toolkit==3.0.29
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.5.3
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rest-condition==1.0.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sentry-sdk==1.5.12
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.3.27
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
watchdog==0.8.3
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.1
